+++ VacationRental.Api.Tests.Integration/Endpoints/Rentals/PutRentalTests.cs
@@ -5,9 +5,10 @@ using System.Threading.Tasks;
 using VacationRental.Api.Contracts.Booking;
 using VacationRental.Api.Contracts.Common;
 using VacationRental.Api.Contracts.Rental;
+using VacationRental.Api.Tests.Integration.Infrastructure;
 using Xunit;
 
-namespace VacationRental.Api.Tests.Integration;
+namespace VacationRental.Api.Tests.Integration.Endpoints.Rentals;
 
 [Collection("Integration")]
 public class PutRentalTests
