+++ VacationRental.Api.Tests.Integration/Infrastructure/IntegrationFixture.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.TestHost;
-using System;
+﻿using System;
 using System.Net.Http;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.TestHost;
 using Xunit;
 
-namespace VacationRental.Api.Tests.Integration
+namespace VacationRental.Api.Tests.Integration.Infrastructure
 {
     [CollectionDefinition("Integration")]
     public sealed class IntegrationFixture : IDisposable, ICollectionFixture<IntegrationFixture>
